+++ lucene/CHANGES.txt
@@ -120,7 +120,9 @@ Other
 
 API Changes
 ---------------------
-(No changes)
+
+* GITHUB#11248: IntBlockPool's SliceReader, SliceWriter, and all int slice functionality are moved out to MemoryIndex.
+  (Stefan Vodita)
 
 New Features
 ---------------------
@@ -151,7 +153,7 @@ Optimizations
 Bug Fixes
 ---------------------
 
-* GITHUB#9660: Throw and ArithmeticException when the offset overflows in a ByteBlockPool. (Stefan Vodita)
+* GITHUB#9660: Throw an ArithmeticException when the offset overflows in a ByteBlockPool. (Stefan Vodita)
 
 * GITHUB#12388: JoinUtil queries were ignoring boosts. (Alan Woodward)
