+++ lucene/core/src/test/org/apache/lucene/index/TestIntBlockPool.java
@@ -16,130 +16,76 @@
  */
 package org.apache.lucene.index;
 
-import java.util.ArrayList;
-import java.util.List;
+import static org.apache.lucene.util.IntBlockPool.INT_BLOCK_SIZE;
+
 import org.apache.lucene.tests.util.LuceneTestCase;
-import org.apache.lucene.util.Counter;
 import org.apache.lucene.util.IntBlockPool;
 
 /** tests basic {@link IntBlockPool} functionality */
 public class TestIntBlockPool extends LuceneTestCase {
+  public void testWriteReadReset() {
+    IntBlockPool pool = new IntBlockPool(new IntBlockPool.DirectAllocator());
+    pool.nextBuffer();
 
-  public void testSingleWriterReader() {
-    Counter bytesUsed = Counter.newCounter();
-    IntBlockPool pool = new IntBlockPool(new ByteTrackingAllocator(bytesUsed));
+    // Write <count> consecutive ints to the buffer, possibly allocating a new buffer
+    int count = random().nextInt(2 * INT_BLOCK_SIZE);
+    for (int i = 0; i < count; i++) {
+      if (pool.intUpto == INT_BLOCK_SIZE) {
+        pool.nextBuffer();
+      }
+      pool.buffer[pool.intUpto++] = i;
+    }
 
-    for (int j = 0; j < 2; j++) {
-      IntBlockPool.SliceWriter writer = new IntBlockPool.SliceWriter(pool);
-      int start = writer.startNewSlice();
-      int num = atLeast(100);
-      for (int i = 0; i < num; i++) {
-        writer.writeInt(i);
-      }
+    // Check that all the ints are present in th buffer pool
+    for (int i = 0; i < count; i++) {
+      assertEquals(i, pool.buffers[i / INT_BLOCK_SIZE][i % INT_BLOCK_SIZE]);
+    }
 
-      int upto = writer.getCurrentOffset();
-      IntBlockPool.SliceReader reader = new IntBlockPool.SliceReader(pool);
-      reader.reset(start, upto);
-      for (int i = 0; i < num; i++) {
-        assertEquals(i, reader.readInt());
-      }
-      assertTrue(reader.endOfSlice());
-      if (random().nextBoolean()) {
-        pool.reset(true, false);
-        assertEquals(0, bytesUsed.get());
-      } else {
-        pool.reset(true, true);
-        assertEquals(IntBlockPool.INT_BLOCK_SIZE * Integer.BYTES, bytesUsed.get());
-      }
+    // Reset without filling with zeros and check that the first buffer still has the ints
+    count = Math.min(count, INT_BLOCK_SIZE);
+    pool.reset(false, true);
+    for (int i = 0; i < count; i++) {
+      assertEquals(i, pool.buffers[0][i]);
+    }
+
+    // Reset and fill with zeros, then check there is no data left
+    pool.intUpto = count;
+    pool.reset();
+    for (int i = 0; i < count; i++) {
+      assertEquals(0, pool.buffers[0][i]);
     }
   }
 
-  public void testMultipleWriterReader() {
-    Counter bytesUsed = Counter.newCounter();
-    IntBlockPool pool = new IntBlockPool(new ByteTrackingAllocator(bytesUsed));
-    for (int j = 0; j < 2; j++) {
-      List<StartEndAndValues> holders = new ArrayList<>();
-      int num = atLeast(4);
-      for (int i = 0; i < num; i++) {
-        holders.add(new StartEndAndValues(random().nextInt(1000)));
-      }
-      IntBlockPool.SliceWriter writer = new IntBlockPool.SliceWriter(pool);
-      IntBlockPool.SliceReader reader = new IntBlockPool.SliceReader(pool);
+  public void testTooManyAllocs() {
+    // Use a mock allocator that doesn't waste memory
+    IntBlockPool pool =
+        new IntBlockPool(
+            new IntBlockPool.Allocator(0) {
+              final int[] buffer = new int[0];
 
-      int numValuesToWrite = atLeast(10000);
-      for (int i = 0; i < numValuesToWrite; i++) {
-        StartEndAndValues values = holders.get(random().nextInt(holders.size()));
-        if (values.valueCount == 0) {
-          values.start = writer.startNewSlice();
-        } else {
-          writer.reset(values.end);
-        }
-        writer.writeInt(values.nextValue());
-        values.end = writer.getCurrentOffset();
-        if (random().nextInt(5) == 0) {
-          // pick one and reader the ints
-          assertReader(reader, holders.get(random().nextInt(holders.size())));
-        }
-      }
+              @Override
+              public void recycleIntBlocks(int[][] blocks, int start, int end) {}
 
-      while (!holders.isEmpty()) {
-        StartEndAndValues values = holders.remove(random().nextInt(holders.size()));
-        assertReader(reader, values);
-      }
-      if (random().nextBoolean()) {
-        pool.reset(true, false);
-        assertEquals(0, bytesUsed.get());
-      } else {
-        pool.reset(true, true);
-        assertEquals(IntBlockPool.INT_BLOCK_SIZE * Integer.BYTES, bytesUsed.get());
+              @Override
+              public int[] getIntBlock() {
+                return buffer;
+              }
+            });
+    pool.nextBuffer();
+
+    boolean throwsException = false;
+    for (int i = 0; i < Integer.MAX_VALUE / INT_BLOCK_SIZE + 1; i++) {
+      try {
+        pool.nextBuffer();
+      } catch (
+          @SuppressWarnings("unused")
+          ArithmeticException ignored) {
+        // The offset overflows on the last attempt to call nextBuffer()
+        throwsException = true;
+        break;
       }
     }
-  }
-
-  private static class ByteTrackingAllocator extends IntBlockPool.Allocator {
-    private final Counter bytesUsed;
-
-    public ByteTrackingAllocator(Counter bytesUsed) {
-      this(IntBlockPool.INT_BLOCK_SIZE, bytesUsed);
-    }
-
-    public ByteTrackingAllocator(int blockSize, Counter bytesUsed) {
-      super(blockSize);
-      this.bytesUsed = bytesUsed;
-    }
-
-    @Override
-    public int[] getIntBlock() {
-      bytesUsed.addAndGet(blockSize * Integer.BYTES);
-      return new int[blockSize];
-    }
-
-    @Override
-    public void recycleIntBlocks(int[][] blocks, int start, int end) {
-      bytesUsed.addAndGet(-((end - start) * blockSize * Integer.BYTES));
-    }
-  }
-
-  private void assertReader(IntBlockPool.SliceReader reader, StartEndAndValues values) {
-    reader.reset(values.start, values.end);
-    for (int i = 0; i < values.valueCount; i++) {
-      assertEquals(values.valueOffset + i, reader.readInt());
-    }
-    assertTrue(reader.endOfSlice());
-  }
-
-  private static class StartEndAndValues {
-    int valueOffset;
-    int valueCount;
-    int start;
-    int end;
-
-    public StartEndAndValues(int valueOffset) {
-      this.valueOffset = valueOffset;
-    }
-
-    public int nextValue() {
-      return valueOffset + valueCount++;
-    }
+    assertTrue(throwsException);
+    assertTrue(pool.intOffset + INT_BLOCK_SIZE < pool.intOffset);
   }
 }
